gathertweets/requirements.txt
flask
tweepy
python-dotenv
numpy
cohere
pandas